+++ Relentless/src/Graphics/Resources/ConstantBuffer.h
@@ -6,10 +6,10 @@ namespace Relentless
 	class ConstantBuffer : public IResource
 	{
 	public:
-		explicit ConstantBuffer(uint32_t sizeInBytes) noexcept;
+		explicit ConstantBuffer(size_t sizeInBytes) noexcept;
 		virtual ~ConstantBuffer() noexcept override final = default;
 
-		uint32_t m_SizeInBytes;
+		size_t m_SizeInBytes;
 		DescriptorHandle m_NonVisibleHandle;
 		DescriptorHandle m_VisibleHandles[3];
 	};
